[minor] fix all the signed/unsigned comparison warnings
I tried to apply the two following principles relatively uniformly:
- if a variable is obviously an array index and nothing more, use size_t
- if a cast in needed anyway, because the loop variable is really used
  as an int, favor signed ints and static_cast<int> the unsigned ones

+++ Goblin Camp/src/Events.cpp
@@ -68,7 +68,7 @@ void Events::Update(bool safe) {
 		else ++immi;
 	}
 
-	if (existingImmigrants.size() < Game::Inst()->OrcCount() / 7 && 
+	if (static_cast<int>(existingImmigrants.size()) < Game::Inst()->OrcCount() / 7 && 
 		Random::Generate(UPDATES_PER_SECOND * 60 * 30) == 0) {
 			SpawnImmigrants();
 	}
